requirements.txt
psycopg2-binary
websockets
requests
pandas
pytz
loguru
keyring
PyJWT
certifi
aiohttp